requirements.txt
# Core Deep Learning Framework
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Transformers and NLP
transformers==4.36.2
tokenizers==0.15.0
datasets==2.16.1
sentencepiece==0.1.99
protobuf==3.20.3

# Language Detection (excluding problematic packages)
langid==1.1.6
# fasttext==0.9.2  # Requires C++ build tools - install separately if needed
# polyglot==16.7.4  # May have compatibility issues
# pycld2==0.41  # May require compilation

# Text Processing
emoji==2.9.0
nltk==3.8.1
regex==2023.12.25
unidecode==1.3.7

# Scientific Computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
scikit-learn==1.3.2

# Configuration and Logging
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Visualization and Monitoring
matplotlib==3.7.5
seaborn==0.13.1
plotly==5.18.0
tqdm==4.66.1
tensorboard==2.15.1

# Model Optimization
onnx==1.15.0
onnxruntime==1.16.3

# API and Interface (for Phase 6)
fastapi==0.109.0
uvicorn==0.27.0
gradio==4.16.0
pydantic==2.5.3

# Testing and Quality
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Utilities
requests==2.31.0
python-json-logger==2.0.7
colorama==0.4.6